feat(test): add test file gatherer, profiler tests, integration tests
Test Files are gathered through a csv listing all the cacphony links and amount of people scanned. The profiler can be used to take a snapshot of the current performance measuring from when found body to measured states. Integration tests are run anytime a change to the algoritm is made to eliminate major breaks.