feat(test): add new cptv-player to test
New player from https://github.com/TheCacophonyProject/cptv-rs fixed a memory issue with last cptv-player that would not dealloc memory. The cptv-rs is a es module which does not play well with ts-node so I have copied a local version & changed imports to "require()" and used process.cwd() for getting dir instead of "import.meta.url" which seems to work fine. Ideally look for a solution at allows for the es module natively