fix(test): memory issues, add reinit in build script, write test file
Wasm requires a number of values to be freed, with the processing only being able to be free'd if a new wasm instance is created, as well as calling cptv player context with free(). This functionality is added within build.sh as it's overwritten each build. Jest was also no correctly awaiting the processing so it wrote only partial profile results.

'