feat(processing): adjust values & conditionals in processing, update wasm-bindgen, add getTemps
This has a lot of logs as tests aren't working for stepping through with debugger, but I ran a number of edge cases to find where it might break and found that "head_is_far_from_edge" would fail quite often, and it could often not find a neck. Removing the conditional & increasing the search area by 10%. There are two other tweaks of to increase some tolerance around head area. It seems to work well with 40% less missed videos and an 18% increase in time to measure. Filtering the hist at x>=2 is likely not needed. Updated wasm to have returned values to be camel instead of underscored e.g. "sample_temp" to "sampleTemp"