fix(processing): find ref only if none found, increase motion threshold by 500
Previous it would return the previous ref if it could not detect a new one. It now returns none if none is found and instead uses the previous ref. This is better logic as the function is trying to detect a new reference. The motion threshold increased was based off a number of videos that would be blurry despite minimal movement