Feat/reliablity changes (#85)
* feat(test): add mocha & chai w/ ts-node, fix linting

Mocha & chai was used to minimize the amount of dependencies required supporting both node & web for the algorithm testing.

* test(analysis): add unit tests & profiling

The FLIR cameras run at 9hz frame rate so we can determine the time it takes from first detection to measured.

* feat(processing): update shell build script

* feat(test): add test file gatherer, profiler tests, integration tests

Test Files are gathered through a csv listing all the cacphony links and amount of people scanned. The profiler can be used to take a snapshot of the current performance measuring from when found body to measured states. Integration tests are run anytime a change to the algoritm is made to eliminate major breaks.

* feat(test): add test file gatherer, profiler tests, integration tests

Test Files are gathered through a csv listing all the cacphony links and amount of people scanned. The profiler can be used to take a snapshot of the current performance measuring from when found body to measured states. Integration tests are run anytime a change to the algoritm is made to eliminate major breaks.

* test: add real temps matching with recordings for profile files

* test(downloader): add ability to get calibration temps

* fix(test): offset value from mem not deallocating

BREAKING CHANGE:

* feat(test): add new cptv-player to test

New player from https://github.com/TheCacophonyProject/cptv-rs fixed a memory issue with last cptv-player that would not dealloc memory. The cptv-rs is a es module which does not play well with ts-node so I have copied a local version & changed imports to "require()" and used process.cwd() for getting dir instead of "import.meta.url" which seems to work fine. Ideally look for a solution at allows for the es module natively

* feat(test): change tests to Jest, add new cptv-player, move cptv-player

moved cptv-player to avoid conflict with new cptv-player (Will implement at later date if needed), added
the https://github.com/TheCacophonyProject/cptv-rs, shifting to jest as mocha start have far too many work
 arounds that jest supported natively. Added a reinit of wasm module, with dealloc as memory seemed to be
shared across multiple files causing bad results.

* fix(test): memory issues, add reinit in build script, write test file

Wasm requires a number of values to be freed, with the processing only being able to be free'd if a new wasm instance is created, as well as calling cptv player context with free(). This functionality is added within build.sh as it's overwritten each build. Jest was also no correctly awaiting the processing so it wrote only partial profile results.

'

* feat(test): add average time/temp to profile_log

* feat(processing): adjust values & conditionals in processing, update wasm-bindgen, add getTemps

This has a lot of logs as tests aren't working for stepping through with debugger, but I ran a number of edge cases to find where it might break and found that "head_is_far_from_edge" would fail quite often, and it could often not find a neck. Removing the conditional & increasing the search area by 10%. There are two other tweaks of to increase some tolerance around head area. It seems to work well with 40% less missed videos and an 18% increase in time to measure. Filtering the hist at x>=2 is likely not needed. Updated wasm to have returned values to be camel instead of underscored e.g. "sample_temp" to "sampleTemp"

* feat(test): add rust cptv player

Not a full implementation

BREAKING CHANGE:

* fix(processing): use wasm-bindgen web instead of no-modules

Requires removing import.meta as it is not supported in Vue 2/vue-cli

* fix(git): would not add submodule so adding to root

* feat(processing): adjust min_face_area, return true on face moved if no prev face, change motion threshold

* fix(processing): remove minor changes that had little significance

While most had 5% increases in terms of speed and relability they were not were not as important asremoving the close to edge conditional & percentage diff of face to prev face which decreased the overall speed by 50% and increased relability by 70%.

* fix(thermal_ref): use previous thermal ref if available

This may be bad logic as if there is always a previous ref and the person it hiding it or it isn't there it would not know, but there are many cases where the ref has not changed (https://browse.cacophony.org.nz/recording/815727) but shows missing ref and using the prev ref is fine. Might be best to say it's missing ref if the measured taken from the ref is below the expect amount in temp

* fix(thermal_ref): missing ref based on abnormal/below average temp

This was due to artifacts of not being able to find the ref despite the reference being clear. While it would be possible to try improve the detection of the circle to remove the random missing ref states, this provides a robust approach of checking the temp of the thermal in the process at it should be constant

* refactor: remove cptv-player

May add at a later date to get fine grain testing.

* fix(processing): find ref only if none found, increase motion threshold by 500

Previous it would return the previous ref if it could not detect a new one. It now returns none if none is found and instead uses the previous ref. This is better logic as the function is trying to detect a new reference. The motion threshold increased was based off a number of videos that would be blurry despite minimal movement

* refactor: minor log cleanup & remove mocha

* feat(test): add false posivtive test & refactor

* feat(processing): reduced min_buffer accumulatin 5 to 2 seconds, reduce area change, reduce dynamic range

The min accumulation would create many artifacts that would be deteced/ghost in the frame, reducing the reset down to 2 seconds seems to reduce the ghosting while improving the accuracy. I previously removed face changed based on area as there were many test files that weren't not measuring because of it, but it dramatically increased false positives, so just had a reduced the area amount

* feat(processing): reduced min_buffer accumulatin 5 to 2 seconds, reduce area change, reduce dynamic range

The min accumulation would create many artifacts that would be deteced/ghost in the frame, reducing the reset down to 2 seconds seems to reduce the ghosting while improving the accuracy. I previously removed face changed based on area as there were many test files that weren't not measuring because of it, but it dramatically increased false positives, so just had a reduced the area amount

* refactor: clean test helper to factory function, minor change to advance state

* fix(processing): adjust face area values & diff area

Problems with merging with the min_buffer so limited the face to be less than 3000. Changed too small 1000 as I saw many viable faces jumping around 1100-1200. Changed diff area face movemont to 40%

* fix(test/process): remove undefined from results, limit head area

Reading when head was far too small, so move condition into face_is_too_small for checking area is greater than certain amount

* fix(test/build): add make build files, fix failed rate on false positive

* fix: limit head height for valid head, minor refactor

* fix(processing): set background to min if large amount of difference in pixel

This dramatically reduces the artifacts being caused by the min_buffer to create a more constant background

* fix: when background hot, don't add to min_buffer

* refactor: rustfmt

* refactor: remove info! & debugger